24.04.2023

+++ Pages/competitions/post.competition.rankings_settings.php
@@ -7,9 +7,10 @@ $rankedCompetitors = db::escape(filter_input(INPUT_POST, 'rankedCompetitors', FI
 $delegates = filter_input(INPUT_POST, 'delegates', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
 $delegates_role = filter_input(INPUT_POST, 'delegates_role', FILTER_DEFAULT, FILTER_REQUIRE_ARRAY);
 $id = $competition->local_id;
-if(!$rankedCompetitors){
-    $rankedCompetitors=100;
+if (!$rankedCompetitors) {
+    $rankedCompetitors = 100;
 }
+
 db::exec("  UPDATE unofficial_competitions
             SET 
                 ranked = $ranked,
@@ -18,25 +19,18 @@ db::exec("  UPDATE unofficial_competitions
                 rankedCompetitors = $rankedCompetitors
             WHERE id = $id ");
 
-
-db::exec(" DELETE FROM  unofficial_competition_delegates where competition_id=$id");
-foreach ($delegates as $j => $delegate) {
-    if ($delegate and $delegates_role[$j] ?? false) {
-        $delegate = db::escape($delegate);
-        if (is_numeric($delegates_role[$j])) {
-            db::exec(" INSERT IGNORE INTO unofficial_competition_delegates (delegate,competition_id,dict_delegate_role) values('$delegate',$id,{$delegates_role[$j]})");
+if ($grand->federation_ext) {
+    db::exec(" DELETE FROM  unofficial_competition_delegates where competition_id=$id");
+    foreach ($delegates as $j => $delegate) {
+        if ($delegate and $delegates_role[$j] ?? false) {
+            $delegate = db::escape($delegate);
+            if (is_numeric($delegates_role[$j])) {
+                db::exec(" INSERT IGNORE INTO unofficial_competition_delegates (delegate,competition_id,dict_delegate_role) values('$delegate',$id,{$delegates_role[$j]})");
+            }
         }
     }
 }
 
 if (!$ranked or!$rankedID) {
     db::exec("UPDATE unofficial_competitions SET rankedID = null, rankedApproved=0, ranked = 0, rankedCompetitors = null WHERE id = $id ");
-    /*db::exec("DELETE from unofficial_competition_delegates WHERE competition_id = $id ");*/
-}
-
-if (!\config::isLocalhost()) {
-    sendMail(
-            config::get('Admin', 'email'), "FunCubing: Rating Competition"
-            , print_r($_POST, true)
-    );
 }
